requirements.txt
wradlib
xarray
numpy
pyyaml
pyproj
scipy
netCDF4
